# apps/infrx-api/openapi/consumer.json
{
  "components": {
    "schemas": {
...
      "Availability": {
        "additionalProperties": false,
        "description": "A feature's availability as FastAPI states it. A failed fetch is `unknown` with a\nreason; it is never rendered as a zero count or a disabled feature.",
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "state": {
            "enum": [
              "configured",
              "disabled",
              "unavailable",
              "unknown"
            ],
            "title": "State",
            "type": "string"
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "required": [
          "state"
        ],
        "title": "Availability",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LabCapabilities": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "features": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Availability"
            },
            "title": "Features",
            "type": "object"
          },
          "provider_org_id": {
            "title": "Provider Org Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "provider_org_id",
          "role",
          "capabilities",
          "features"
        ],
        "title": "LabCapabilities",
        "type": "object"
      },
...
      "ListPage_Workspace_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Workspace"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "data"
        ],
        "title": "ListPage[Workspace]",
        "type": "object"
      },
...
      "MemberGrant": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "maxLength": 320,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "role": {
            "enum": [
              "viewer",
              "developer",
              "administrator"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "email",
          "role"
        ],
        "title": "MemberGrant",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "Workspace": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "provider_name": {
            "title": "Provider Name",
            "type": "string"
          },
          "provider_org_id": {
            "title": "Provider Org Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "provider_org_id",
          "provider_name",
          "role",
          "capabilities"
        ],
        "title": "Workspace",
        "type": "object"
      },
...
      "infrx__contracts__api__ListPage_Member___1": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/infrx__state__identity__Member"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "data"
        ],
        "title": "ListPage[Member]",
        "type": "object"
      },
...
      "infrx__state__identity__Member": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "granted_at": {
            "format": "date-time",
            "title": "Granted At",
            "type": "string"
          },
          "granted_by": {
            "title": "Granted By",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "email",
          "role",
          "granted_by",
          "granted_at"
        ],
        "title": "Member",
        "type": "object"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "bearerFormat": "sk-infrx",
        "description": "A consumer or provider-dev API key (`/v1/*`).",
        "scheme": "bearer",
        "type": "http"
      },
      "CheckpointSignature": {
        "description": "HMAC of the raw body by a LAB_CHECKPOINT_KEYS key.",
        "in": "header",
        "name": "X-Infrx-Signature",
        "type": "apiKey"
      },
      "OperatorKey": {
        "bearerFormat": "sk-infrx",
        "description": "An operator-audience API key.",
        "scheme": "bearer",
        "type": "http"
      },
      "SessionBearer": {
        "bearerFormat": "JWT",
        "description": "The signed-in user's Supabase session, forwarded by the web app's server; verified by the auth server.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "infrx consumer",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/lab/v1/capabilities": {
      "get": {
        "operationId": "getLabV1Capabilities",
        "parameters": [
          {
            "in": "query",
            "name": "provider_org_id",
            "required": false,
            "schema": {
              "default": "",
              "title": "Provider Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabCapabilities"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Lab Capabilities"
      }
    },
...
    "/lab/v1/workspaces": {
      "get": {
        "operationId": "getLabV1Workspaces",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPage_Workspace_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "List Workspaces"
      }
    },
    "/lab/v1/workspaces/{provider_org_id}/members": {
      "get": {
        "operationId": "getLabV1WorkspacesByProviderOrgIdMembers",
        "parameters": [
          {
            "in": "path",
            "name": "provider_org_id",
            "required": true,
            "schema": {
              "title": "Provider Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/infrx__contracts__api__ListPage_Member___1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "List Members"
      },
      "post": {
        "operationId": "postLabV1WorkspacesByProviderOrgIdMembers",
        "parameters": [
          {
            "in": "path",
            "name": "provider_org_id",
            "required": true,
            "schema": {
              "title": "Provider Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberGrant"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/infrx__state__identity__Member"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Grant"
      }
    },
    "/lab/v1/workspaces/{provider_org_id}/members/{user_id}": {
      "delete": {
        "operationId": "deleteLabV1WorkspacesByProviderOrgIdMembersByUserId",
        "parameters": [
          {
            "in": "path",
            "name": "provider_org_id",
            "required": true,
            "schema": {
              "title": "Provider Org Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/infrx__state__identity__Member"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Revoke"
      }
    },
...
  }
}